+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ New Features
 - Multispectral: NBU satellite image dataset, ERGAS+SAM metrics, PanNet, generalised pansharpening and decolorize (:gh:`371` by `Julian Tachella`_ and `Andrew Wang`_)
 - StackedPhysics: class definition, loss and data-fidelity (:gh:`371` by `Julian Tachella`_ and `Andrew Wang`_)
 - Added HyperSpectral Unmixing operator (:gh:`353` by `Dongdong Chen`_ and `Andrew Wang`_)
+- Added new diffusion sampler DEFT (:gh:`378` by `Alexander Denker`_)
 
 - Add validation dataset to data generator (:gh:`363` by `Andrew Wang`_)
 - Add Rescale and ToComplex torchvision-style transforms (:gh:`363` by `Andrew Wang`_)